requirements.txt
fastapi==0.109.1
uvicorn==0.15.0
pydantic==1.10.13
python-jose==3.3.0
python-dotenv==0.19.0
PyJWT==2.4.0
httpx==0.23.0
aiohttp==3.9.4
pytest-asyncio==0.16.0
coverage==7.6.1